requirements.txt
google-auth-oauthlib==1.2.0
google-auth==2.27.0
google-api-python-client==1.12.5
slack-bolt==1.18.1
openai==1.3.0
tavily-python==0.0.1
python-dotenv==1.0.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.13.0
aiohttp==3.9.1
